Enhance chat feature and improve shop functionalities

+++ src/main/java/com/project/web_hand_made_TMDT/model/ChatMessage.java
@@ -20,7 +20,7 @@ public class ChatMessage {
     private ChatRoom chatRoom;
 
     @Column(name = "sender_id")
-    private Long senderId;
+    private int senderId;
 
     @Column(name = "sender_name")
     private String senderName;
